+++ app/src/main/java/com/audiolearning/app/data/db/entities/BaseEntity.kt
@@ -1,5 +1,5 @@
 package com.audiolearning.app.data.db.entities
 
 interface BaseEntity {
-    var id: Long?
+    var id: Int?
 }
